Create README.md
Assignment 

Week 5 of my Assignment on zuri front-end web development internship program, 
consisting of three branches, which are mainly week 3, 4 and 5 tasks. 
And this is basically the week 5 tasks in a different branch.
## 🚀 About Me
I'm an aspiring front-end web developer currently enrolled in an internship with zuri training. 


## 🛠 Skills
Javascript, HTML, CSS...


## Other Common Github Profile Sections
👩‍💻 I'm currently working on building projects with html css and javascript...

🧠 I'm currently learning front-end web development...